updated rendering so assets use 'maze' space coordinates, maze fills screen

+++ src/maze.hpp
@@ -21,7 +21,7 @@ public:
     virtual void draw(sf::RenderTarget& target, sf::RenderStates states) const;
     Cell getCell(int row, int column) const;
 
-    void setPlayerPosition(sf::Vector2i pos);
+    void setPlayerPosition(sf::Vector2f pos); // in row/column space
     void setCursorPosition(sf::Vector2i pos);
     sf::Vector2i extendFromCursor(sf::Vector2i p) const;
 
@@ -29,6 +29,6 @@ public:
     const int MAZE_HEIGHT;
 private:
     std::set<std::string> _openWalls;
-    sf::Vector2i _playerPosition;
+    sf::Vector2f _playerPosition;
     sf::Vector2i _cursorPosition;
 };
